requirements.txt
mycroft-skills-sdk==0.1.20
python-todotxt

